+++ Mods/S7_Journal_6b1aee95-bd4c-4190-a919-4aa7096eaa2b/Story/RawFiles/Lua/BootstrapClient.lua
@@ -3,6 +3,7 @@
 --  =======
 
 Ext.Require("Shared/Auxiliary.lua")
+Ext.Require("Client/ContextMenu.lua")
 Ext.Require("Client/TreasureTables.lua")
 
 --  ==================