+++ Mods/S7_Journal_6b1aee95-bd4c-4190-a919-4aa7096eaa2b/Story/RawFiles/Lua/BootstrapServer.lua
@@ -3,7 +3,9 @@
 --  ======
 
 Ext.Require("Shared/Auxiliary.lua")
-if Ext.IsDeveloperMode() then Ext.Require("Server/DevMode.lua") end
+Ext.Require("Server/ContextMenuHandler.lua")
+
+if Ext.IsDeveloperMode() then Ext.Require("Server/Development/DevMode.lua") end
 
 --  =================
 --  GAME START EVENTS