+++ Mods/S7_Journal_6b1aee95-bd4c-4190-a919-4aa7096eaa2b/Story/RawFiles/Lua/Server/ContextMenuHandler.lua
@@ -1,0 +1,8 @@
+--  ====================
+--  CONTEXT MENU HANDLER
+--  ====================
+
+Ext.RegisterNetListener("S7UCL_ContextMenu", function(channel, payload)
+    if payload == "271" then S7Debug:Print("271 SUCCESS") end
+    if payload == "272" then S7Debug:Print("272 SUCCESS") end
+end)